# Add method to validate Shopify webhooks using `ReadOnlyMemory<byte>`