--- a/ShopifySharp.Tests/GenerateGraphQLSchema_Test.cs
+++ b/ShopifySharp.Tests/GenerateGraphQLSchema_Test.cs
@@ -1,4 +1,4 @@
-#if NET7_0_OR_GREATER
+#if NET8_0_OR_GREATER
 using System.Collections.Generic;
 using System.IO;
 using System.Text;